+++ online-clothing-store-backend/src/main/java/com/app/repository/WishlistRepository.java
@@ -3,6 +3,6 @@ package com.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.app.pojos.Wishlist;
 
-public interface WishlistRepository extends JpaRepository<Wishlist,Long> {
+public interface WishlistRepository extends JpaRepository<Wishlist,Integer> {
 
 }
